this is .pkl extension file from jupyter notebook.
Car Price Prediction Model
This repository contains a machine learning model for predicting car prices based on features such as year, model, and mileage. The project uses a scikit-learn pipeline to streamline data preprocessing and model training, and the trained model is saved in a .pkl file for easy deployment and reuse.